Cleaning up operators/transforms.
This is purely code motion which is splitting out bits into seperate
files
